only admins can delete

+++ backend/routes/users.js
@@ -17,7 +17,7 @@ router.post("/", async (req, res) => {
     const salt = await bcrypt.genSalt(Number(process.env.SALT));
     const hashPassword = await bcrypt.hash(req.body.password, salt);
 
-    await new User({ ...req.body, password: hashPassword }).save();
+    await new User({ ...req.body, password: hashPassword, role: "customer" }).save();
     res.status(201).send({ message: "User created successfully" });
   } catch (error) {
     res.status(500).send({ message: "Internal server error!!!" });
@@ -37,6 +37,13 @@ router.get("/", async (req, res) => {
 router.delete("/:id", async (req, res) => {
   try {
     const { id } = req.params;
+    const user = await User.findById(id);
+    if (!user) {
+      return res.status(404).send({message: "User not found"});
+    }
+    if (user.role !== "admin") {
+      return res.status(401).send({message: "Unauthorized operation"});
+    }
     await User.deleteOne({ _id: id });
     res.send("Got a DELETE request at /users");
   } catch (error) {
